+++ src/models/NewVirtualMachineRequiredVariables.ts
@@ -1,0 +1,4 @@
+export interface NewVirtualMachineRequiredVariables {
+  key: string;
+  value: string;
+}