+++ src/models/vagrant/VagrantCloudBoxes.ts
@@ -1,0 +1,5 @@
+import {VagrantCloudBox} from "./VagrantCloudBox";
+
+export interface VagrantCloudBoxes {
+  boxes: VagrantCloudBox[];
+}